LocationSelector functions moved
in this commit I was able to move functions related only to the LocationSelector in the right file in order to make the code cleaner
